+++ src/main/java/com/sayildiz/kbe/calculator/exception/TooManyDecimalsException.java
@@ -1,7 +1,9 @@
 package com.sayildiz.kbe.calculator.exception;
 
+import java.math.BigDecimal;
+
 public class TooManyDecimalsException extends RuntimeException {
-    public TooManyDecimalsException(double price){
+    public TooManyDecimalsException(BigDecimal price){
         super("Value should have maximal two decimals " + price);
     }
 }
